requirements.txt
discord.py
colorthief
colormap
easydev
Levenshtein
